CentralDisplay.py::Changes to InputForm Behaviour.
